bdal: Add a block device abstraction layer
bdal: Add a block device abstraction layer

Add a basic block device abstraction layer. This
abstracts the interface between Linux AIO and the
SPDK NVMe driver. It only provides the common
intersection of features between the two.

There are a number of missing pieces of functionality
in this implementation, so it is not suitable for
real use quite yet. This serves mostly as the
starting point for review and discussion.

Change-Id: Iab2fc782d4e3c92e2f961cc9b3bc5220887c92e9
Signed-off-by: Ben Walker <benjamin.walker@intel.com>